src/requirements.txt
flask
flask-sqlalchemy
pymysql
python-slugify
flask-admin
